Removing extra tests from DataObjectTests.ClipboardTests class
